Fisher Information computation and plots.
Script doing a few computations and plot to compute the Fisher Information, according to separable 2D Von Mises tuning curves, and assuming a Gaussian approximation to the recall process. First assume no covariance structure in the FI, then use the measured one.

We will then need to use the theoretical covariance derived using a KL projection (~Laplace) approximation for the whole recall model.

This contributes to issue #10, doesn't really close it yet, the whole thing has to be improved.