Interesting discrepancy between sampler and fisher information
I found that in the sampler output, it is actually possible to get a power
law exponent different from 1. It seems to be controlled by rc_scale, sigma_x
didn't have much input there (though it does depend on the number of samples quite drastically).

This actually is or is not a good thing depending on the explanation:
- It means the Fisher information is a bad approximation for now, or that there is some mistake somewhere
- It means that we can create more complicated power laws, arising when sampling during the inference
process, which is good. Also should mean that even with a linear model, sampling and rc_scale can
introduce effects that could falsely indicate a non-unity power law exponent.

(work on #17 and #15, possibly #16)