Plots and works for the Fisher Info optimization
Ugly early version of a parameter optimisation and sweeper for the Fisher Information.

Checked the optimal sigma_x and kappa. Used either a least square solution or fitting a power law and matching those statistics. Tried then to get the coupled optimal value, and this seem to depend quite a lot on the score function.

- Sigma_x: No control over power law exponent (~1), big nonlinear/exponential dependance on the y0 of the power law (and similarly, on the highest value of the Fisher Information, i.e. the precision when one object is present).
- Kappa: Very small control over power law exponent (still stays around 1 though). Inverse effect than sigma_x on fisher information.
- Both: when using the least square, good overall fit and corresponds to the optimal parameters to the sampler (TO CHECK). When using the power law, completely different ones (crazy sigma_x = 16), but may be due to missing sources of noise in the current FI computation.

(work on #10, #15, #17)