Data loading made nicer
Improved the way data from PBS is loaded back in. Can now give short data description dictionaries, and a few scripts will load everything behind the scenes.
Handles multiple parameters, multiple dimensions and construct numpy arrays (of unknown size, yay) automatically.

Should then add a function to look for the parameters in the dataset themselves, not in the filenames, and it should be quite robust. The dictionary should be saved with the data, to be even cleaner.
Work on #11