DataIO created
Created this class to handle saving and reloading (to implement later) of results arrays. Handles unique filenames nicely.

Just needs to pass it a list of variables to save to a file, and the locals() dictionary (slightly ugly, but at least we define the list of variables only once per script and in a nice compact fashion).