Important bug in precision computation
This explains the missing square relationship between the rcscale and the fisher information estimated through the precision. This could actually make the simulations and the theory agree now, but this remains to be verified.
I'd rather reimplement the whole thing in 2D again to avoid loosing time tracing errors in the framework.

Work on #17, #15