Bug fixed Fisher information from posterior curvature
- Fixed: there was a bug in the posterior estimation and the fisher information computation. Now fits the theory very well.

This now doesn't correspond to the fisher information estimated from the precision. There seem to be a problem with the samples generated, they nearly alway overestimate the standard deviation. Hence more noise is in the system and we fail to achieve the Cramer-Rao bound. Hard to know what's causing it, still need to investigate.

The good fit between the curvature and the theory happens when the curvature uses a diagonal covariance (just like the model used in the theory, obviously). For the full covariance, the match is good for small sigma, but disapear afterwards.

work on #17, #10