New Sampler using Inverse Transform Sampling
Way faster than Slice sampling... not sure why we didn't do that before.
Actually I know, it's because we didn't want to integrate the likelihood
to find the normalising function, but I've been doing it quite a lot already,
so let's just do it and not worry.

The speedups are crazy to sample... But the overhead is quite high when first
estimating the inverse CDF, so need to be careful for large N.